ready for move to the ubuntu server

+++ libraries/DataFlash/NavigationRecord.cpp
@@ -151,9 +151,9 @@ bool NavigationMSG::_parse_navigation()
 			_roll = _buffer.data.roll;
 			_pitch = _buffer.data.pitch;
 			_yaw = _buffer.data.yaw;
-			_vx = (float)_buffer.data.vx / 100.0f;
-			_vy = (float)_buffer.data.vy / 100.0f;
-			_vz = (float)_buffer.data.vz / 100.0f;
+			_vx = _buffer.data.vx / 100.0f;
+			_vy = _buffer.data.vy / 100.0f;
+			_vz = _buffer.data.vz / 100.0f;
 			_x = _buffer.data.px / 100.0;
 			_y = _buffer.data.py / 100.0;
 			_z = _buffer.data.pz / 100.0;
